Remove old test files, rename new test files (scripts not yet changed)
